update kdocs and file structure
