+++ agents/agents-features/agents-features-event-handler/src/jvmTest/kotlin/ai/koog/agents/features/eventHandler/feature/ModelResolvingTestExecutor.kt
@@ -24,7 +24,7 @@ class ModelResolvingTestExecutor(
 ) : PromptExecutor() {
 
     override suspend fun resolveModel(
-        requestedModel: LLModel,
+        model: LLModel,
         promptExecutorOperation: PromptExecutorOperation,
     ): ResolvedModel = ResolvedModel(effectiveModel)
