+++ prompt/prompt-executor/prompt-executor-model/src/commonMain/kotlin/ai/koog/prompt/executor/llms/RoutingLLMPromptExecutor.kt
@@ -124,16 +124,16 @@ public open class RoutingLLMPromptExecutor @JvmOverloads constructor(
      * @throws ModelResolutionException If no client is found for the model's provider and no fallback is configured.
      */
     override suspend fun resolveModel(
-        requestedModel: LLModel,
+        model: LLModel,
         promptExecutorOperation: PromptExecutorOperation
     ): ResolvedModel = when {
-        requestedModel.provider in supportedProviders -> ResolvedModel(effectiveModel = requestedModel)
+        model.provider in supportedProviders -> ResolvedModel(effectiveModel = model)
 
         fallback != null -> ResolvedModel(effectiveModel = fallback.fallbackModel)
 
         else -> throw ModelResolutionException(
-            requestedModel,
-            "No client found for provider: ${requestedModel.provider}"
+            model,
+            "No client found for provider: ${model.provider}"
         )
     }
