+++ prompt/prompt-executor/prompt-executor-model/src/commonMain/kotlin/ai/koog/prompt/executor/model/ResolvedModel.kt
@@ -3,19 +3,9 @@ package ai.koog.prompt.executor.model
 import ai.koog.prompt.llm.LLModel
 
 /**
- * The model selected by [PromptExecutor.resolveModel] for execution
+ * Represents a resolved model for use by [PromptExecutorAPI]. Outcome of [PromptExecutorAPI.resolveModel]
+ *
+ * @property effectiveModel The effective [LLModel] instance that will be used by [PromptExecutorAPI]
+ * for performing operations.
  */
 public data class ResolvedModel(public val effectiveModel: LLModel)
-
-public enum class PromptExecutorOperation {
-    Execute,
-    Moderate,
-    MultipleChoices,
-    Streaming
-}
-
-public class ModelResolutionException(
-    requestedModel: LLModel,
-    detailedMsg: String? = null,
-    cause: Throwable? = null
-) : Exception("Model resolution failed for requested model: $requestedModel." + detailedMsg?.let { "\n\t$it" }, cause)
